Automatically generate the meetings lists
Switch to using Liquid templates for the homepage and move the actual
meeting information into a Jekyll data file.

At the moment, this isn't hugely different to the static version but
with the data and presentation separated, we can hopefully start to do
more interesting things (particularly if we attach more metadata to the
meetings such as books discussed).