Add Amazon affiliate links to books we've read
Addresses issue #6.

I've added these links to the books we've read. They were generated using the
Amazon affiliate tools. We can probably auto-generate them and add them
automatically if we move the book lists into `_data/`, but we can do
that at a later stage.

For the record, the affiliate account is administered by me, Chris Lowis, but
I commit to using any revenues to purchase books to donate to the London
Computation Club for the use of any members.