Redesign the layout of the site
Gone is our beloved cellular automaton, replaced with a photograph from
one of our meetings (we could choose another which is better
representation of the meetings).

Simplify fonts to Roboto for headings and the best system sans-serif we
can find.

With the automaton gone, remove the "rule"-based copy in favour of an
distinct code of conduct section and re-order the sections on the
homepage in terms of likely interest.