Fix event timings in the calendar
As we only specify the date of a meeting in the site data, we hard coded
the start and end time as 18:30 and 20:30. This works fine when the
London timezone matches UTC but, alas, Daylight Savings Time means that
meetings will be an hour out at certain points in the year.

As we're restricted in the amount of logic we can put in Jekyll
templates, we take our meeting date (which is parsed as a Europe/London
time thanks to the timezone setting in our site configuration) and now
add 18:30 and 19:30 in seconds. This hopefully means that the start and
end times of events will adjust to DST appropriately.